% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_availability.R
\name{data_availability}
\alias{data_availability}
\title{Data availability Function}
\usage{
data_availability(codes, met_directory)
}
\arguments{
\item{codes}{List of metcodes that you want to check the availability for specific variables e.g. wind speed}

\item{met_directory}{Input met_directory defined by the user - assumes the data in csv format}
}
\description{
This function allows you to check the availability of met stations across a given region and time frame.
Useful if you want to combine met data with AQ data.
Note, this function is dependent on having the met data downloaded on a local system.
}
\examples{
data_availability()
}
\keyword{met_data_availability}
