% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_aq_combined.R
\name{met_aq_combined}
\alias{met_aq_combined}
\title{Met AQ combined function
Should you want to run either the deweather or rmweather function, this code will allow you create a dataframe with
the required sites, pollutants and will match the AQ site to the closest met site. Note, the met_directory should be
downloaded on a local system.}
\usage{
met_aq_combined(met_directory, begin_date, end_date, aqcode, metcode)
}
\arguments{
\item{met_directory}{Input met_directory defined by the user - assumes the data in csv format}

\item{begin_date}{Beginning date in yyyy-mm-dd string format}

\item{end_date}{End date in yyyy-mm-dd string format}

\item{aqcode}{List of metcodes based on AURN, KCL etc. using openair}

\item{metcode}{List of met sites based on importNOAA}
}
\description{
Met AQ combined function
Should you want to run either the deweather or rmweather function, this code will allow you create a dataframe with
the required sites, pollutants and will match the AQ site to the closest met site. Note, the met_directory should be
downloaded on a local system.
}
\examples{
met_uk_df_all = purrr::map2_dfr(.x = aurn_noaa_nearest_urb_sites$code, .y = aurn_noaa_nearest_urb_sites$met_code, .f = ~read_met_sites(directory_met_data, "2006-01-01", "2011-12-31", aqcode = .x, metcode = .y))
met_aq_combined()
}
\keyword{met_aq_combined}
