% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_aq_prepared_rm_vs_de.R
\name{met_aq_prepared_rm_vs_de}
\alias{met_aq_prepared_rm_vs_de}
\title{Met AQ combined prepared rm vs de function
Should you want to run either the deweather or rmweather function, this code will allow you create a dataframe with
the required sites and prepare it to be normalised for a single pollutant choice}
\usage{
met_aq_prepared_rm_vs_de(df, pollutant_type, rmweather = FALSE)
}
\arguments{
\item{df}{Output of combined Met and AQ dataset}

\item{pollutant_type}{Input for pollutant to test code - only been tested for one pollutant (tested for nox)}

\item{rmweather}{Default at FALSE. If set to TRUE, dataframe will be suitable for rmweather package. If FALSE,
dataframe will be suitable for deweather}
}
\description{
Met AQ combined prepared rm vs de function
Should you want to run either the deweather or rmweather function, this code will allow you create a dataframe with
the required sites and prepare it to be normalised for a single pollutant choice
}
\examples{
met_aq_prepared_rm_vs_de()

}
\keyword{met_aq_combined}
