% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_normalised_data_comparison.R
\name{pull_normalised_data_comparison}
\alias{pull_normalised_data_comparison}
\title{Pull normalised data comparison function
Should you want to run either the deweather or rmweather function, this code will allow you create a dataframe with
the required sites and prepare it to be normalised for a single pollutant choice}
\usage{
pull_normalised_data_comparison(df_de, df2_rm, df3_raw, pollutant_type, site)
}
\arguments{
\item{df_de}{Output of normalised pollutants with deweather module}

\item{df3_raw}{Output of raw pollutants}

\item{pollutant_type}{Input for pollutant to test code - only been tested for one pollutant (tested for nox)}

\item{site}{List of sites used in normalisation code}

\item{df_rm}{Output of normalised pollutants with rmweather module}
}
\description{
Pull normalised data comparison function
Should you want to run either the deweather or rmweather function, this code will allow you create a dataframe with
the required sites and prepare it to be normalised for a single pollutant choice
}
\examples{
pull_normalised_data_comparison()
}
\keyword{pull_normalised_statistics}
