% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_normalised_statistics.R
\name{pull_normalised_statistics}
\alias{pull_normalised_statistics}
\title{Pull normalised statistics function}
\usage{
pull_normalised_statistics(df_de, df2_rm, site)
}
\arguments{
\item{df_de}{Output of normalised pollutants with deweather module}

\item{site}{List of sites used in normalisation code}

\item{df_rm}{Output of normalised pollutants with rmweather module}
}
\description{
Should you want to run either the deweather or rmweather function, this code will allow you create a dataframe with
the required sites and prepare it to be normalised for a single pollutant choice
}
\examples{
pull_normalised_statistics()
}
\keyword{pull_normalised_statistics}
